refactor(utils): add utils to shared folder
