+++ src/core/shared/utils/generate-uuid/index.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from 'uuid';
 
 export const generateUUID = (): string => {
-	return uuidv4();
+  return uuidv4();
 };
