+++ src/core/shared/utils/hash/index.ts
@@ -1,13 +1,13 @@
 import { compare, hash } from 'bcryptjs';
 
 export const hashPassword = async (password: string): Promise<string> => {
-	const saltRounds = 8;
-	return hash(password, saltRounds);
+  const saltRounds = 8;
+  return hash(password, saltRounds);
 };
 
 export const comparePassword = async (
-	password: string,
-	hash: string,
+  password: string,
+  hash: string,
 ): Promise<boolean> => {
-	return compare(password, hash);
+  return compare(password, hash);
 };
