--- src/shared/utils/generate-uuid/index.spec.ts
@@ -1,11 +1,0 @@
-import { generateUUID } from './';
-
-describe('Generate UUID', () => {
-	it('should generate a UUID', async () => {
-		const uuid = generateUUID();
-
-		expect(uuid).toBeDefined();
-		expect(typeof uuid).toBe('string');
-		expect(uuid.length).toBe(36);
-	});
-});
